Renaming wait_free_unordered_map to unordered_map
